+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # v0.1.6 2020-06-18
 
-- Added `NxtSupport::AssignableValues`
+- Added `NxtSupport::PreprocessAttributes`
 
 # v0.1.5 2020-06-17